src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
cachecontrol==0.14.0
    # via firebase-admin
cachetools==5.4.0
    # via google-auth
certifi==2024.7.4
    # via requests
cffi==1.17.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==43.0.0
    # via pyjwt
django==5.1
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
firebase-admin==6.5.0
    # via -r requirements.in
google-api-core[grpc]==2.19.1
    # via
    #   firebase-admin
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-api-python-client==2.141.0
    # via firebase-admin
google-auth==2.33.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-cloud-core==2.4.1
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.17.2
    # via firebase-admin
google-cloud-storage==2.18.2
    # via firebase-admin
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.63.2
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.65.4
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.65.4
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.7
    # via requests
msgpack==1.0.8
    # via cachecontrol
proto-plus==1.24.0
    # via
    #   google-api-core
    #   google-cloud-firestore
protobuf==5.27.3
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pycparser==2.22
    # via cffi
pyjwt[crypto]==2.9.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   firebase-admin
pyparsing==3.1.2
    # via httplib2
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via
    #   cachecontrol
    #   google-api-core
    #   google-cloud-storage
rsa==4.9
    # via google-auth
sqlparse==0.5.1
    # via django
typing-extensions==4.12.2
    # via asgiref
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.2
    # via requests
